Categoric filters in scatter plot
